% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-log.R
\name{ingest_default_query}
\alias{ingest_default_query}
\title{Default PubMed query for pharmacometric literature}
\usage{
ingest_default_query()
}
\value{
Character scalar PubMed query string.
}
\description{
Default PubMed query for pharmacometric literature
}
