% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-entrez.R
\name{ingest_entrez_search}
\alias{ingest_entrez_search}
\title{Search PubMed via E-utilities}
\usage{
ingest_entrez_search(query, retmax = 20L, cfg)
}
\arguments{
\item{query}{PubMed query string.}

\item{retmax}{Maximum PMIDs to return.}

\item{cfg}{Ingest config from \code{\link[=ingest_load_config]{ingest_load_config()}}.}
}
\value{
Character vector of PMIDs.
}
\description{
Search PubMed via E-utilities
}
