% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-gui-job.R
\name{ingest_gui_worker}
\alias{ingest_gui_worker}
\title{Background worker for LibeRary GUI jobs}
\usage{
ingest_gui_worker(params)
}
\arguments{
\item{params}{List with job settings (see Shiny app).}
}
\value{
List with \code{ok} and result or error.
}
\description{
Background worker for LibeRary GUI jobs
}
