% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-log.R
\name{ingest_latest_manifest}
\alias{ingest_latest_manifest}
\title{Find most recent discover manifest}
\usage{
ingest_latest_manifest(cfg)
}
\arguments{
\item{cfg}{Ingest config list.}
}
\value{
File path or \code{""}.
}
\description{
Find most recent discover manifest
}
