% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-log.R
\name{ingest_make_logger}
\alias{ingest_make_logger}
\title{Create a file-backed logger for GUI / background jobs}
\usage{
ingest_make_logger(log_path)
}
\arguments{
\item{log_path}{Path to append log lines.}
}
\value{
Function \verb{function(msg, level = "INFO")}.
}
\description{
Create a file-backed logger for GUI / background jobs
}
