% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-oa.R
\name{ingest_pmc_oa_lookup}
\alias{ingest_pmc_oa_lookup}
\title{Check PMC open-access availability and PDF link}
\usage{
ingest_pmc_oa_lookup(pmcid, cfg)
}
\arguments{
\item{pmcid}{PMC ID with or without prefix.}

\item{cfg}{Ingest config.}
}
\value{
List with is_oa, pdf_url.
}
\description{
Check PMC open-access availability and PDF link
}
