% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-manifest.R
\name{ingest_read_manifest}
\alias{ingest_read_manifest}
\title{Read manifest CSV}
\usage{
ingest_read_manifest(path)
}
\arguments{
\item{path}{CSV path.}
}
\value{
Data frame.
}
\description{
Read manifest CSV
}
