% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-log.R
\name{ingest_read_progress}
\alias{ingest_read_progress}
\title{Read GUI progress state}
\usage{
ingest_read_progress(progress_path)
}
\arguments{
\item{progress_path}{JSON file path.}
}
\value{
List or NULL.
}
\description{
Read GUI progress state
}
