% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest-config.R
\name{ingest_validate_config}
\alias{ingest_validate_config}
\title{Validate and normalize LibeRary configuration}
\usage{
ingest_validate_config(cfg)
}
\arguments{
\item{cfg}{Configuration list.}
}
\value{
Normalized configuration.
}
\description{
Validate and normalize LibeRary configuration
}
