% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical-qualification.R
\name{library_clinical_qualification_validate}
\alias{library_clinical_qualification_validate}
\title{Validate a clinical-use qualification}
\usage{
library_clinical_qualification_validate(qualification)
}
\arguments{
\item{qualification}{Qualification record.}
}
\value{
Validated qualification.
}
\description{
Validate a clinical-use qualification
}
